feat: 월별 불가능한 인원수 및 닉네임 반환 추가

+++ backend/src/main/java/com/whenwemeet/backend/domain/meetingRoom/dto/response/MeetingRoomInfoResponse.java
@@ -16,7 +16,7 @@ import java.util.List;
 public class MeetingRoomInfoResponse {
 
     private String name;
-    private Long memberNumber;
+    private Integer memberNumber;
     private LocalDateTime meetingDate;
     private List<UserInfoResponse> info;
